Refactor import path checks and enhance decision-making logic
- Updated the `is_relative` function to only consider paths starting with "./" or "../" as relative.
- Introduced `SuppressReason` enum in `decision.rs` to categorize suppression reasons for rule decisions.
- Modified the `decide` function to utilize `SuppressReason` for clearer suppression messages.
- Added new tests for TypeScript imports to ensure absolute paths and scoped packages are skipped.
- Implemented a new `sanitize` module for code quality audits, providing functions to strip comments and string literals from code lines.
- Created a pipeline for file, framework, and project audits, including registration and execution of various audits.
- Ensured all registered audits have associated rule metadata and are unique within their scopes.
