+++ src/graph/imports/common.rs
@@ -14,5 +14,5 @@ pub(super) fn extract_string_literal(s: &str) -> Option<&str> {
 }
 
 pub(super) fn is_relative(path: &str) -> bool {
-    path.starts_with('.') || path.starts_with('/')
+    path.starts_with("./") || path.starts_with("../")
 }
