+++ src/knowledge/decision.rs
@@ -7,6 +7,43 @@ use crate::knowledge::model::{RuleDecision, RuleDecisionAction, RuleMatchContext
 use crate::scan::facts::{FileFacts, ScanFacts};
 use crate::scan::path_classification::is_low_signal_audit_path;
 use std::collections::HashSet;
+use std::fmt;
+
+#[derive(Debug, Clone, Copy, PartialEq, Eq)]
+enum SuppressReason {
+    LanguageNotMatched,
+    LanguageSupportTooLow,
+    FrameworkNotMatched,
+    RuntimeNotMatched,
+    ParadigmNotMatched,
+    LowSignalPath,
+    ConfigFile,
+    GeneratedFile,
+}
+
+impl fmt::Display for SuppressReason {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        let s = match self {
+            SuppressReason::LanguageNotMatched => "rule does not apply to this language",
+            SuppressReason::LanguageSupportTooLow => {
+                "language support level is below this rule's minimum"
+            }
+            SuppressReason::FrameworkNotMatched => "rule does not apply to this framework",
+            SuppressReason::RuntimeNotMatched => "rule does not apply to this runtime",
+            SuppressReason::ParadigmNotMatched => "rule does not apply to this paradigm",
+            SuppressReason::LowSignalPath => "low-signal audit path",
+            SuppressReason::ConfigFile => "configuration file",
+            SuppressReason::GeneratedFile => "generated file",
+        };
+        f.write_str(s)
+    }
+}
+
+impl From<SuppressReason> for String {
+    fn from(reason: SuppressReason) -> Self {
+        reason.to_string()
+    }
+}
 
 pub fn decide(context: &RuleMatchContext<'_>) -> RuleDecision {
     let Some(rule) = bundled_knowledge().rule_by_id(context.rule_id) else {
@@ -19,13 +56,13 @@ pub fn decide(context: &RuleMatchContext<'_>) -> RuleDecision {
             .iter()
             .any(|l| rule.languages.contains(*l))
     {
-        return RuleDecision::suppress("rule does not apply to this language");
+        return RuleDecision::suppress(SuppressReason::LanguageNotMatched);
     }
 
     if let Some(minimum_support) = rule.minimum_support
         && !language_support_satisfies(rule, context.languages, minimum_support)
     {
-        return RuleDecision::suppress("language support level is below this rule's minimum");
+        return RuleDecision::suppress(SuppressReason::LanguageSupportTooLow);
     }
 
     if !rule.frameworks.is_empty()
@@ -34,11 +71,11 @@ pub fn decide(context: &RuleMatchContext<'_>) -> RuleDecision {
             .iter()
             .any(|f| rule.frameworks.contains(*f))
     {
-        return RuleDecision::suppress("rule does not apply to this framework");
+        return RuleDecision::suppress(SuppressReason::FrameworkNotMatched);
     }
 
     if !rule.runtimes.is_empty() && !context.runtimes.iter().any(|r| rule.runtimes.contains(*r)) {
-        return RuleDecision::suppress("rule does not apply to this runtime");
+        return RuleDecision::suppress(SuppressReason::RuntimeNotMatched);
     }
 
     if !rule.paradigms.is_empty()
@@ -47,19 +84,19 @@ pub fn decide(context: &RuleMatchContext<'_>) -> RuleDecision {
             .iter()
             .any(|p| rule.paradigms.contains(*p))
     {
-        return RuleDecision::suppress("rule does not apply to this paradigm");
+        return RuleDecision::suppress(SuppressReason::ParadigmNotMatched);
     }
 
     if context.is_low_signal && rule.suppress_low_signal {
-        return RuleDecision::suppress("low-signal audit path");
+        return RuleDecision::suppress(SuppressReason::LowSignalPath);
     }
 
     if rule.suppress_config && context.roles.contains(&"config") {
-        return RuleDecision::suppress("configuration file");
+        return RuleDecision::suppress(SuppressReason::ConfigFile);
     }
 
     if rule.suppress_generated && context.roles.contains(&"generated") {
-        return RuleDecision::suppress("generated file");
+        return RuleDecision::suppress(SuppressReason::GeneratedFile);
     }
 
     let mut decision = RuleDecision::apply(context.base_severity);
@@ -85,24 +122,8 @@ pub fn decide_for_audit_context(
     base_severity: Severity,
     signal: Option<&str>,
 ) -> RuleDecision {
-    let framework_ids = context.framework_ids();
-    let role_ids = context.role_ids();
-    let paradigm_ids = context.paradigm_ids();
-    let runtime_ids = context.runtime_ids();
     let language_ids = [context.language_id()];
-
-    decide(&RuleMatchContext {
-        rule_id,
-        languages: &language_ids,
-        frameworks: &framework_ids,
-        roles: &role_ids,
-        paradigms: &paradigm_ids,
-        runtimes: &runtime_ids,
-        is_test: context.is_test,
-        is_low_signal: false,
-        signal,
-        base_severity,
-    })
+    decide_with_context(rule_id, &language_ids, context, false, base_severity, signal)
 }
 
 pub fn decide_for_file(
@@ -114,6 +135,18 @@ pub fn decide_for_file(
     let context = classify_file(file);
     let mut language_ids = language_ids_for_file(file, &context);
     dedup_static_ids(&mut language_ids);
+    let is_low_signal = is_low_signal_audit_path(&file.path) && !context.is_test;
+    decide_with_context(rule_id, &language_ids, &context, is_low_signal, base_severity, signal)
+}
+
+fn decide_with_context(
+    rule_id: &str,
+    language_ids: &[&str],
+    context: &AuditContext,
+    is_low_signal: bool,
+    base_severity: Severity,
+    signal: Option<&str>,
+) -> RuleDecision {
     let framework_ids = context.framework_ids();
     let role_ids = context.role_ids();
     let paradigm_ids = context.paradigm_ids();
@@ -121,13 +154,13 @@ pub fn decide_for_file(
 
     decide(&RuleMatchContext {
         rule_id,
-        languages: &language_ids,
+        languages: language_ids,
         frameworks: &framework_ids,
         roles: &role_ids,
         paradigms: &paradigm_ids,
         runtimes: &runtime_ids,
         is_test: context.is_test,
-        is_low_signal: is_low_signal_audit_path(&file.path) && !context.is_test,
+        is_low_signal,
         signal,
         base_severity,
     })
